# raptor: real second pass with reader