--- a/source/routing/raptor_solution_reader.cpp
+++ b/source/routing/raptor_solution_reader.cpp
@@ -409,7 +409,7 @@ Solutions read_solutions(const RAPTOR& raptor,
             } catch (stop_search&) {}
         }
     }
-    return reader.solutions;
+    return std::move(reader.solutions);
 }
 
 } // anonymous namespace