--- a/source/routing/raptor_solution_reader.h
+++ b/source/routing/raptor_solution_reader.h
@@ -54,13 +54,13 @@ struct Journey {
     bool better_on_sn(const Journey& that, bool) const;
     friend std::ostream& operator<<(std::ostream& os, const Journey& j);
 
-    std::vector<Section> sections;
+    std::vector<Section> sections;// the pt sections, with transfer between them
     navitia::time_duration sn_dur = 0_s;// street network duration
     navitia::time_duration transfer_dur = 0_s;// walking duration during transfer
     navitia::time_duration min_waiting_dur = 0_s;// minimal waiting duration on every transfers
     DateTime departure_dt = 0;// the departure dt of the journey, including sn
     DateTime arrival_dt = 0;// the arrival dt of the journey, including sn
-    uint8_t nb_vj_extentions = 0;
+    uint8_t nb_vj_extentions = 0;// number of vehicle journey extentions (I love useless comments!)
 };
 
 // this structure compare 2 solutions.  It chooses which solutions