With good bounds on best_labels.

Performances:
 - fr-pdl: 60s -> 54.1s => 10% faster (0 times stop after 100000 solutions)
 - idf: 48s -> 40.3s => 16% faster (2 times stop after 100000 solutions)
 - israel: 100s -> 85.9s => 14% faster (22 times stop after 100000 solution)